+++ app/api/notification/[id]/route.tsx
@@ -1,5 +1,7 @@
 import prisma from "@/prisma/client";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
+import { authOptions } from "../../auth/authOptions";
 
 interface Props{
     params:{
@@ -17,6 +19,8 @@ export async function GET(request : NextRequest , {params : {id}} : Props){
 }
 
 export async function DELETE(request : NextRequest, {params} : Props){
+    const session = await getServerSession(authOptions);
+    if (!session) return NextResponse.json({err : "Unauthroized"},{status:401})
     const data = await prisma.notification.delete({
         where:{
             id : parseInt(params.id)
